refactor: protectedroute перемещен в hoc
