+++ packages/client/src/hoc/protected-route/index.tsx
@@ -1,13 +1,16 @@
 import { FC } from 'react'
 import { Navigate } from 'react-router-dom'
-import { useAuth } from '../auth-status/index'
+
 import { Page } from '../../utils/constants/navigation'
+import { useAuth } from '../../hooks/auth-status'
 
 interface ProtectedRouteProps {
   element: JSX.Element
 }
 
-export const ProtectedRoute: FC<ProtectedRouteProps> = ({ element }) => {
+const ProtectedRoute: FC<ProtectedRouteProps> = ({ element }) => {
   const { isAuth } = useAuth()
   return isAuth ? element : <Navigate to={`/${Page.SIGNIN}`} replace />
 }
+
+export default ProtectedRoute
